Add Datasplit.py

+++ model.py
@@ -55,7 +55,7 @@ def unet3D(pretrained_weights = None,input_size = (132,132,132,1), stage2=True):
     conv9 = Conv3D(32, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(merge9)
     conv9 = Conv3D(32, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv9)
     conv9 = Conv3D(2, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv9)
-    conv10 = Conv3D(1, 1, activation = 'sigmoid')(conv9)
+    conv10 = Conv3D(1, 1, activation = 'sigmoid', dtype='float32')(conv9)
 
     model = Model(inputs = inputs, outputs = conv10)
